requirements.txt
anthropic>=0.25.0
sympy>=1.12
numpy>=1.26
streamlit>=1.35.0
python-dotenv>=1.0.0
